+++ core/transaction/bundle_tx.py
@@ -4,6 +4,7 @@ import os
 
 from core.transaction.transaction import Transaction
 from core.configuration.configuration import Configuration
+from core.wallet.wallet import get_private_key, get_public_key
 
 def bundle_tnx(size, reward_amount):
     """
@@ -19,23 +20,28 @@ def bundle_tnx(size, reward_amount):
         with open('{0}/verified_transactions.json'.format(os.path.join(os.getcwd(), r'data')), 'r') as file:
             data = json.load(file)
             file.close()
-
-            # TODO: bundle as many transactions as possible
-            tx_temp = []
-            tx_temp.append(data.popitem())
-            tx_temp.append(data.popitem())
-            block_transactions.update(tx_temp)
+    except IOError:
+        print("~~~~~~~~~~~~~~~~")
         with open('{0}/verified_transactions.json'.format(os.path.join(os.getcwd(), r'data')), 'w') as file:
+            data = {}
             json.dump(data, file)
             file.close()
-    except IOError as e:
-        # file does not exist or not able to read file
-        print('{0}'.format(e))
+
+    try:  # TODO: bundle as many transactions as possible
+        tx_temp = []
+        tx_temp.append(data.popitem())
+        tx_temp.append(data.popitem())
+        block_transactions.update(tx_temp)
     except KeyError as e:
         # there was not at least 2 transactions in verified_transactions.json
         print('Need at least 2 transaction per block\n{0}'.format(e))
 
+    with open('{0}/verified_transactions.json'.format(os.path.join(os.getcwd(), r'data')), 'w') as file:
+        json.dump(data, file)
+        file.close()
+
     return block_transactions
+
 
 def create_coinbase_tx(reward_amount):
     """
@@ -45,5 +51,5 @@ def create_coinbase_tx(reward_amount):
     """
     cbtx = Transaction()
     config = Configuration()
-    cbtx.add_coinbase_output(config.get_conf("key").get("public"), reward_amount)
+    cbtx.add_coinbase_output(get_public_key("string"), reward_amount)
     return cbtx
